requirements.txt
aiohttp
asyncio
discord
discord-components
discord.py
discord-py-slash-command
requests
pillow